+++ upnp/src/genlib/net/sock.c
@@ -2,6 +2,7 @@
  *
  * Copyright (c) 2000-2003 Intel Corporation 
  * All rights reserved. 
+ * Copyright (c) 2012 France Telecom All rights reserved. 
  *
  * Redistribution and use in source and binary forms, with or without 
  * modification, are permitted provided that the following conditions are met: 
@@ -112,7 +113,7 @@ static int sock_read_write(
 	/*! [out] Buffer to get data to or send data from. */
 	char *buffer,
 	/*! [in] Size of the buffer. */
-	int bufsize,
+	size_t bufsize,
 	/*! [in] timeout value. */
 	int *timeoutSecs,
 	/*! [in] Boolean value specifying read or write option. */
@@ -166,7 +167,7 @@ static int sock_read_write(
 #endif
 		if (bRead) {
 			/* read data. */
-			numBytes = (long)recv(sockfd, buffer, (size_t)bufsize, MSG_NOSIGNAL);
+			numBytes = (long)recv(sockfd, buffer, bufsize, MSG_NOSIGNAL);
 		} else {
 			byte_left = bufsize;
 			bytes_sent = 0;
@@ -200,12 +201,12 @@ static int sock_read_write(
 	return (int)numBytes;
 }
 
-int sock_read(SOCKINFO *info, char *buffer, int bufsize, int *timeoutSecs)
+int sock_read(SOCKINFO *info, char *buffer, size_t bufsize, int *timeoutSecs)
 {
 	return sock_read_write(info, buffer, bufsize, timeoutSecs, TRUE);
 }
 
-int sock_write(SOCKINFO *info, const char *buffer, int bufsize, int *timeoutSecs)
+int sock_write(SOCKINFO *info, const char *buffer, size_t bufsize, int *timeoutSecs)
 {
 	/* Consciently removing constness. */
 	return sock_read_write(info, (char *)buffer, bufsize, timeoutSecs, FALSE);
