+++ upnp/src/inc/config.h
@@ -2,6 +2,7 @@
  *
  * Copyright (c) 2000-2003 Intel Corporation
  * All rights reserved.
+ * Copyright (c) 2012 France Telecom All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -189,7 +190,7 @@
  *
  * @{
  */
-#define WEB_SERVER_BUF_SIZE  (1024*1024)
+#define WEB_SERVER_BUF_SIZE  (size_t)(1024*1024)
 /* @} */
 
 /*!
